Minor refactor in Contact interface
