+++ CHT/library/api/fermat-cht-api/src/main/java/com/bitdubai/fermat_cht_api/layer/middleware/interfaces/Contact.java
@@ -2,7 +2,7 @@ package com.bitdubai.fermat_cht_api.layer.middleware.interfaces;
 
 import com.bitdubai.fermat_api.layer.all_definition.components.enums.PlatformComponentType;
 
-import java.sql.Date;
+import java.util.Date;
 import java.util.UUID;
 
 /**
